Extract stubs to separate namespace
