+++ stubs/StubAdapter.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Queue\Adapter;
+namespace Yiisoft\Queue\Stubs;
 
+use Yiisoft\Queue\Adapter\AdapterInterface;
 use Yiisoft\Queue\Enum\JobStatus;
 use Yiisoft\Queue\Message\MessageInterface;
