+++ stubs/StubWorker.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Queue\Worker;
+namespace Yiisoft\Queue\Stubs;
 
 use Yiisoft\Queue\Message\MessageInterface;
 use Yiisoft\Queue\QueueInterface;
+use Yiisoft\Queue\Worker\WorkerInterface;
 
 /**
  * Stub worker that does nothing.
