+++ tests/Unit/Stubs/StubQueueTest.php
@@ -2,14 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Queue\Tests\Unit;
+namespace Yiisoft\Queue\Tests\Unit\Stubs;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Queue\Adapter\StubAdapter;
 use Yiisoft\Queue\Message\Message;
-use Yiisoft\Queue\Queue;
 use Yiisoft\Queue\QueueInterface;
-use Yiisoft\Queue\StubQueue;
+use Yiisoft\Queue\Stubs\StubQueue;
+use Yiisoft\Queue\Stubs\StubAdapter;
 
 final class StubQueueTest extends TestCase
 {
